+++ DataAccess/BookstoreDatabaseSettings.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DataAccess
+{
+    public class BookstoreDatabaseSettings : IBookstoreDatabaseSettings
+    {
+        public string BooksCollectionName { get; set; }
+        public string ConnectionString { get; set; }
+        public string DatabaseName { get; set; }
+    }
+
+    public interface IBookstoreDatabaseSettings
+    {
+        string BooksCollectionName { get; set; }
+        string ConnectionString { get; set; }
+        string DatabaseName { get; set; }
+    }
+}
