+++ DataAccess/DataAccess.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="MongoDB.Driver.Core" Version="2.12.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Task1.Core\Task1.Core.csproj" />
+  </ItemGroup>
+
 </Project>
