+++ DataAccess/Repositories/BookRepository.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Text;
+using Task1.Core.Models;
+using Task1.Core.Repositories;
+
+namespace DataAccess.Repositories
+{
+    public class BookRepository : IBookRepository
+    {
+        public void Add(Book book)
+        {
+            throw new NotImplementedException();
+        }
+
+        public void Edit(Book book)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Book FindById(string Id)
+        {
+            throw new NotImplementedException();
+        }
+
+        public IEnumerable GetProducts()
+        {
+            throw new NotImplementedException();
+        }
+
+        public void Remove(string Id)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
