+++ StoreApi/Models/Book.cs
@@ -1,25 +1,25 @@
-﻿using MongoDB.Bson;
-using MongoDB.Bson.Serialization.Attributes;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿//using MongoDB.Bson;
+//using MongoDB.Bson.Serialization.Attributes;
+//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Threading.Tasks;
 
-namespace StoreApi.Models
-{
-    public class Book
-    {
-        [BsonId]
-        [BsonRepresentation(BsonType.ObjectId)]
-        public string Id { get; set; }
+//namespace StoreApi.Models
+//{
+//    public class Book
+//    {
+//        [BsonId]
+//        [BsonRepresentation(BsonType.ObjectId)]
+//        public string Id { get; set; }
 
-        [BsonElement("Name")]
-        public string BookName { get; set; }
+//        [BsonElement("Name")]
+//        public string BookName { get; set; }
 
-        public decimal Price { get; set; }
+//        public decimal Price { get; set; }
 
-        public string Category { get; set; }
+//        public string Category { get; set; }
 
-        public string Author { get; set; }
-    }
-}
+//        public string Author { get; set; }
+//    }
+//}
