+++ StoreApi/Models/BookstoreDatabaseSettings.cs
@@ -1,21 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Threading.Tasks;
 
-namespace StoreApi.Models
-{
-    public class BookstoreDatabaseSettings : IBookstoreDatabaseSettings
-    {
-        public string BooksCollectionName { get; set; }
-        public string ConnectionString { get; set; }
-        public string DatabaseName { get; set; }
-    }
+//namespace StoreApi.Models
+//{
+//    public class BookstoreDatabaseSettings : IBookstoreDatabaseSettings
+//    {
+//        public string BooksCollectionName { get; set; }
+//        public string ConnectionString { get; set; }
+//        public string DatabaseName { get; set; }
+//    }
 
-    public interface IBookstoreDatabaseSettings
-    {
-        string BooksCollectionName { get; set; }
-        string ConnectionString { get; set; }
-        string DatabaseName { get; set; }
-    }
-}
+//    public interface IBookstoreDatabaseSettings
+//    {
+//        string BooksCollectionName { get; set; }
+//        string ConnectionString { get; set; }
+//        string DatabaseName { get; set; }
+//    }
+//}
