+++ StoreApi/Services/BookService.cs
@@ -1,44 +1,45 @@
-﻿using MongoDB.Driver;
-using StoreApi.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿//using MongoDB.Driver;
+//using StoreApi.Models;
+//using System;
+//using System.Collections.Generic;
+//using System.Linq;
+//using System.Threading.Tasks;
+//using Task1.Core.Models;
 
-namespace StoreApi.Services
-{
-    public class BookService
-    {
-        private readonly IMongoCollection<Book> _books;
+//namespace StoreApi.Services
+//{
+//    public class BookService
+//    {
+//        private readonly IMongoCollection<Book> _books;
 
-        public BookService(IBookstoreDatabaseSettings settings)
-        {
+//        public BookService(IBookstoreDatabaseSettings settings)
+//        {
 
-            var client = new MongoClient(settings.ConnectionString);
-            var database = client.GetDatabase(settings.DatabaseName);
+//            var client = new MongoClient(settings.ConnectionString);
+//            var database = client.GetDatabase(settings.DatabaseName);
 
-            _books = database.GetCollection<Book>(settings.BooksCollectionName);
-        }
+//            _books = database.GetCollection<Book>(settings.BooksCollectionName);
+//        }
 
-        public List<Book> Get() =>
-            _books.Find(book => true).ToList();
+//        public List<Book> Get() =>
+//            _books.Find(book => true).ToList();
 
-        public Book Get(string id) =>
-            _books.Find<Book>(book => book.Id == id).FirstOrDefault();
+//        public Book Get(string id) =>
+//            _books.Find<Book>(book => book.Id == id).FirstOrDefault();
 
-        public Book Create(Book book)
-        {
-            _books.InsertOne(book);
-            return book;
-        }
+//        public Book Create(Book book)
+//        {
+//            _books.InsertOne(book);
+//            return book;
+//        }
 
-        public void Update(string id, Book bookIn) =>
-            _books.ReplaceOne(book => book.Id == id, bookIn);
+//        public void Update(string id, Book bookIn) =>
+//            _books.ReplaceOne(book => book.Id == id, bookIn);
 
-        public void Remove(Book bookIn) =>
-            _books.DeleteOne(book => book.Id == bookIn.Id);
+//        public void Remove(Book bookIn) =>
+//            _books.DeleteOne(book => book.Id == bookIn.Id);
 
-        public void Remove(string id) =>
-            _books.DeleteOne(book => book.Id == id);
-    }
-}
+//        public void Remove(string id) =>
+//            _books.DeleteOne(book => book.Id == id);
+//    }
+//}
