+++ StoreApi/Startup.cs
@@ -7,12 +7,13 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using StoreApi.Models;
-using StoreApi.Services;
+using Task1.Core;
+using DataAccess;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using DataAccess.DataServices;
 
 namespace StoreApi
 {
