+++ Task1.Core/Models/Book.cs
@@ -1,0 +1,23 @@
+﻿using MongoDB.Bson;
+using MongoDB.Bson.Serialization.Attributes;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Task1.Core.Models
+{    public class Book
+    {
+        [BsonId]
+        [BsonRepresentation(BsonType.ObjectId)]
+        public string Id { get; set; }
+
+        [BsonElement("Name")]
+        public string BookName { get; set; }
+
+        public decimal Price { get; set; }
+
+        public string Category { get; set; }
+
+        public string Author { get; set; }
+    }
+}
