+++ Task1.Core/Task1.Core.csproj
@@ -1,0 +1,10 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="MongoDB.Driver" Version="2.12.3" />
+		<PackageReference Include="MongoDB.Driver.Core" Version="2.12.3" />
+	</ItemGroup>
+</Project>
